Revision 1 Version 9.6
Fix a bug for Event object (this fault insert in previous revision)
Add more control on objects of type Group.
We can add members (variables, arrays, functions and modules), but these
additions holded if a group made in current module or function.
Try { } now preserve stack for subs/if/For-Next/gosub label